Allow thereafter in sampler to be zero
When creating a new Sampler, there is nothing stopping us from passing
in 0 as a value for thereafter. One might expect this to be legal -- as
the documentation doesn't say thereafter should be greater than zero and
no error is returned when passing in this configuration -- and to mean
"drop all messages after the first n". What actually happens is a

    panic: runtime error: integer divide by zero

when we try to log a message.

This patch avoids the division by zero and enforces the behavior
described above.